update the actiondemo
