--- unittests/actiondemo/actiondemo.abi
@@ -1,119 +1,0 @@
-{
-  "____comment": "This file was generated by eosio-abigen. DO NOT EDIT - 2019-01-07T10:42:22",
-  "version": "eosio::abi/1.0",
-  "types": [],
-  "structs": [{
-      "name": "seedobj",
-      "base": "",
-      "fields": [{
-          "name": "id",
-          "type": "uint64"
-        },{
-          "name": "create",
-          "type": "time_point"
-        },{
-          "name": "seedstr",
-          "type": "string"
-        },{
-          "name": "txid",
-          "type": "string"
-        },{
-          "name": "action",
-          "type": "uint64"
-        }
-      ]
-    },{
-      "name": "args",
-      "base": "",
-      "fields": [{
-          "name": "loop",
-          "type": "uint64"
-        },{
-          "name": "num",
-          "type": "uint64"
-        }
-      ]
-    },{
-      "name": "generate",
-      "base": "",
-      "fields": [{
-          "name": "t",
-          "type": "args"
-        }
-      ]
-    },{
-      "name": "clear",
-      "base": "",
-      "fields": []
-    },{
-      "name": "args_name",
-      "base": "",
-      "fields": [{
-          "name": "name",
-          "type": "name"
-        }
-      ]
-    },{
-      "name": "hascontract",
-      "base": "",
-      "fields": [{
-          "name": "t",
-          "type": "args_name"
-        }
-      ]
-    },{
-      "name": "args_inline",
-      "base": "",
-      "fields": [{
-          "name": "payer",
-          "type": "name"
-        },{
-          "name": "in",
-          "type": "name"
-        }
-      ]
-    },{
-      "name": "inlineact",
-      "base": "",
-      "fields": [{
-          "name": "t",
-          "type": "args_inline"
-        }
-      ]
-    }
-  ],
-  "actions": [{
-      "name": "generate",
-      "type": "generate",
-      "ricardian_contract": ""
-    },{
-      "name": "clear",
-      "type": "clear",
-      "ricardian_contract": ""
-    },{
-      "name": "hascontract",
-      "type": "hascontract",
-      "ricardian_contract": ""
-    },{
-      "name": "inlineact",
-      "type": "inlineact",
-      "ricardian_contract": ""
-    }
-  ],
-  "tables": [{
-      "name": "seedobjs",
-      "index_type": "i64",
-      "key_names": [
-        "id"
-      ],
-      "key_types": [
-        "uint64"
-      ],
-      "type": "seedobj"
-    }
-  ],
-  "ricardian_clauses": [],
-  "error_messages": [],
-  "abi_extensions": [],
-  "variants": []
-}
