+++ unittests/actiondemo/actiondemo/actiondemo.abi
@@ -1,0 +1,159 @@
+{
+    "____comment": "This file was generated with eosio-abigen. DO NOT EDIT Fri Apr 26 17:07:28 2019",
+    "version": "eosio::abi/1.1",
+    "structs": [
+        {
+            "name": "apply",
+            "base": "",
+            "fields": [
+                {
+                    "name": "contract",
+                    "type": "name"
+                },
+                {
+                    "name": "act",
+                    "type": "name"
+                }
+            ]
+        },
+        {
+            "name": "args",
+            "base": "",
+            "fields": [
+                {
+                    "name": "loop",
+                    "type": "uint64"
+                },
+                {
+                    "name": "num",
+                    "type": "uint64"
+                }
+            ]
+        },
+        {
+            "name": "args_name",
+            "base": "",
+            "fields": [
+                {
+                    "name": "name",
+                    "type": "name"
+                }
+            ]
+        },
+        {
+            "name": "argsinline",
+            "base": "",
+            "fields": [
+                {
+                    "name": "payer",
+                    "type": "name"
+                },
+                {
+                    "name": "in",
+                    "type": "name"
+                }
+            ]
+        },
+        {
+            "name": "clear",
+            "base": "",
+            "fields": []
+        },
+        {
+            "name": "generate",
+            "base": "",
+            "fields": [
+                {
+                    "name": "t",
+                    "type": "args"
+                }
+            ]
+        },
+        {
+            "name": "hascontract",
+            "base": "",
+            "fields": [
+                {
+                    "name": "t",
+                    "type": "args_name"
+                }
+            ]
+        },
+        {
+            "name": "inlineact",
+            "base": "",
+            "fields": [
+                {
+                    "name": "t",
+                    "type": "argsinline"
+                }
+            ]
+        },
+        {
+            "name": "seedobj",
+            "base": "",
+            "fields": [
+                {
+                    "name": "id",
+                    "type": "uint64"
+                },
+                {
+                    "name": "create",
+                    "type": "time_point"
+                },
+                {
+                    "name": "seedstr",
+                    "type": "string"
+                },
+                {
+                    "name": "txid",
+                    "type": "string"
+                },
+                {
+                    "name": "action",
+                    "type": "uint64"
+                }
+            ]
+        }
+    ],
+    "types": [],
+    "actions": [
+        {
+            "name": "apply",
+            "type": "apply",
+            "ricardian_contract": ""
+        },
+        {
+            "name": "clear",
+            "type": "clear",
+            "ricardian_contract": ""
+        },
+        {
+            "name": "generate",
+            "type": "generate",
+            "ricardian_contract": ""
+        },
+        {
+            "name": "hascontract",
+            "type": "hascontract",
+            "ricardian_contract": ""
+        },
+        {
+            "name": "inlineact",
+            "type": "inlineact",
+            "ricardian_contract": ""
+        }
+    ],
+    "tables": [
+        {
+            "name": "seedobjs",
+            "type": "seedobj",
+            "index_type": "i64",
+            "key_names": [],
+            "key_types": []
+        }
+    ],
+    "ricardian_clauses": [],
+    "variants": [],
+    "abi_extensions": []
+}
