+++ unittests/actiondemo/wallet.txt
@@ -1,0 +1,1 @@
+PW5KBvkVfkUHqp7S7oDeesvPbk4BMe4pwDtN14HygMHxYiTLftrve
